Changed test structure; added flask db script
